Prepping for a big change to playermovement, also: Made some slight changes to code and set up a rabbit that will later be animated, killable, and cookable!
I'm going to be changing the way the player interacts with the environment, using raycasting will prove far less cumbersome, more accurate, and will require less lines of code to change in the future. I predict this change will shave the lines of code in the playermovement by half. This upload is to make sure I have a solid build from before I attempt this in case I break something and need to try again.